# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: LEGO Factory v3 API
  description: |
    REST API for the LEGO Factory v3 Manufacturing Platform.

    ## Overview
    This API provides endpoints for:
    - **SCADA**: Machine control, tags, alarms, historian
    - **MES**: Work orders, scheduling, OEE, labor tracking
    - **ERP**: Sales orders, purchasing, inventory, financial
    - **QMS**: Documents, NCR/CAPA, training, calibration
    - **CMMS**: Assets, maintenance, work orders
    - **LEGO**: Brick designs, slicing, catalog
    - **ML**: Inference, predictions, anomaly detection
    - **Unity**: Digital twin state synchronization
    - **ROS2**: Robot control and cell orchestration

    ## Authentication
    Currently supports session-based authentication via /api/auth endpoints.

    ## Rate Limiting
    API endpoints are not rate-limited in the current version.

  version: 3.0.0
  contact:
    name: LEGO Factory Support
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://lego-factory:5000
    description: Docker container
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
...